% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_categories.R
\name{categories}
\alias{categories}
\title{Filter by subject categories}
\usage{
categories(obj, ...)
}
\arguments{
\item{obj}{A scielo_query object.}

\item{...}{Subject categories to filter by.}
}
\value{
The modified query object.
}
\description{
Filter by subject categories
}
\examples{
query <- search_scielo("ecology") |> categories("environmental sciences")
}
