% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_journal.R
\name{journals}
\alias{journals}
\title{Filter by journal names}
\usage{
journals(obj, ...)
}
\arguments{
\item{obj}{A scielo_query object.}

\item{...}{Journal names to filter by.}
}
\value{
The modified query object.
}
\description{
Filter by journal names
}
\examples{
query <- search_scielo("biology") |> journals("Revista de Biología Tropical")
}
