% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_languages.R
\name{languages}
\alias{languages}
\title{Filter by languages}
\usage{
languages(obj, ..., operator = "OR")
}
\arguments{
\item{obj}{A scielo_query object.}

\item{...}{Language codes to filter by ("es", "pt", "en").}

\item{operator}{Boolean operator ("AND" or "OR").}
}
\value{
The modified query object.
}
\description{
Filter by languages
}
\examples{
query <- search_scielo("medicine") |> languages("es", "en")
}
