% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_nmax.R
\name{nmax}
\alias{nmax}
\title{Set the maximum number of rows to return}
\usage{
nmax(obj, value)
}
\arguments{
\item{obj}{A scielo_query object.}

\item{value}{An integer number.}
}
\value{
The modified query object.
}
\description{
Set the maximum number of rows to return
}
