% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_collections.R
\name{normalize_collections}
\alias{normalize_collections}
\title{Normalize SciELO collection names or ISO codes}
\usage{
normalize_collections(collections)
}
\arguments{
\item{collections}{A character vector of length 1: a country name (e.g., "Costa Rica")
or a valid SciELO ISO code (e.g., "cri").}
}
\value{
A character string representing the normalized SciELO collection code.
}
\description{
Converts country names or ISO codes into valid SciELO collection codes.
Only one value is allowed.
}
\examples{
normalize_collections("Costa Rica")  # returns "cri"
normalize_collections("cri")         # returns "cri"

}
