% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_nmax.R
\name{normalize_nmax}
\alias{normalize_nmax}
\title{Normalize and validate n_max}
\usage{
normalize_nmax(value)
}
\arguments{
\item{value}{The value to validate.}
}
\value{
An integer or NULL.
}
\description{
Ensures that the value of n_max is a positive integer or NULL.
}
\keyword{internal}
