% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.scielo_query.R
\name{print.scielo_query}
\alias{print.scielo_query}
\title{Print method for scielo_query objects}
\usage{
\method{print}{scielo_query}(x, ...)
}
\arguments{
\item{x}{A \code{scielo_query} object created with \code{search_scielo()}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \code{data.frame} containing the query results (title, authors, year, DOI, abstract).
}
\description{
Executes the query stored in a \code{scielo_query} object and fetches the results as a data frame.
When the object is printed or evaluated, it will display the search results.
}
\examples{
\dontrun{
# Search for articles and get results as a data frame
query_results <- search_scielo("environmental sciences") |>
   journals("Uniciencia") |>
   categories("sciences") |>
   languages("es") |>
   nmax(10)

# The 'query_results' object is a data.frame and will be printed automatically
query_results

# You can also explicitly print it or assign it to a variable
my_df <- print(query_results)
head(my_df)
}
}
