% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_scielo.R
\name{search_scielo}
\alias{search_scielo}
\title{Create a new SciELO query object}
\usage{
search_scielo(query, lang = "es")
}
\arguments{
\item{query}{Search term.}

\item{lang}{Language (default "es").}
}
\value{
An object of class \code{scielo_query}.
}
\description{
Create a new SciELO query object
}
